# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api's Clinica Recoleta",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://endpoint.centrooftalmorecoleta.net"
    },
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
...
    "/api/horario/{horarioId}": {
      "get": {
        "tags": [
          "Horario"
        ],
        "summary": "Obtener Horario por ID",
        "description": "Retorna la información de un Horario por su ID",
        "operationId": "getHorarioById",
        "parameters": [
          {
            "name": "horarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profesional_id": {
                      "type": "number"
                    },
                    "dia": {
                      "type": "string"
                    },
                    "desde": {
                      "type": "string"
                    },
                    "hasta": {
                      "type": "string"
                    },
                    "tiempo": {
                      "type": "string"
                    },
                    "usuario": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Horario no encontrado"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/api/horario/date": {
      "get": {
        "tags": [
          "Horario"
        ],
        "summary": "Obtener Horario por Fecha",
        "description": "Retorna los Horarios para una fecha específica",
        "operationId": "getHorarioByDate",
        "parameters": [
          {
            "name": "fecha",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "profesional_id": {
                        "type": "number"
                      },
                      "dia": {
                        "type": "string"
                      },
                      "desde": {
                        "type": "string"
                      },
                      "hasta": {
                        "type": "string"
                      },
                      "tiempo": {
                        "type": "string"
                      },
                      "usuario": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron Horarios para la fecha especificada"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/api/horario/dateprofesional/{idProfesional}": {
      "get": {
        "tags": [
          "Horario"
        ],
        "summary": "Obtener Horario por ID de Profesional",
        "description": "Retorna los Horarios para un Profesional específico",
        "operationId": "getHorarioByProfesionalId",
        "parameters": [
          {
            "name": "idProfesional",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "profesional_id": {
                        "type": "number"
                      },
                      "dia": {
                        "type": "string"
                      },
                      "desde": {
                        "type": "string"
                      },
                      "hasta": {
                        "type": "string"
                      },
                      "tiempo": {
                        "type": "string"
                      },
                      "usuario": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron Horarios para el Profesional especificado"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/api/horario/createHorario": {
      "post": {
        "tags": [
          "Horario"
        ],
        "summary": "Crear Horario",
        "description": "Crea un nuevo Horario",
        "operationId": "createHorario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "profesional_id": {
                    "type": "number"
                  },
                  "dia": {
                    "type": "string"
                  },
                  "desde": {
                    "type": "string"
                  },
                  "hasta": {
                    "type": "string"
                  },
                  "tiempo": {
                    "type": "string"
                  },
                  "usuario": {
                    "type": "string"
                  }
                },
                "required": [
                  "profesional_id",
                  "dia",
                  "desde",
                  "hasta",
                  "tiempo",
                  "usuario"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Horario creado exitosamente"
          },
          "400": {
            "description": "Solicitud inválida, faltan datos o son incorrectos"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/api/horario/deleteHorario/{horarioId}": {
      "delete": {
        "tags": [
          "Horario"
        ],
        "summary": "Eliminar Horario por ID",
        "description": "Elimina un Horario por su ID",
        "operationId": "deleteHorario",
        "parameters": [
          {
            "name": "horarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Horario eliminado exitosamente"
          },
          "404": {
            "description": "Horario no encontrado"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiAuth": {
        "type": "http",
        "description": "Login with email and password to get the authentication token",
        "name": "Token based Based",
        "in": "header",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  }
}